rightsize/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#

plotly==1.10.0
psutil==4.2.0
python-magic==0.4.11
pytz==2016.4              # via plotly
pyyaml==3.11
redis==2.10.5
requests==2.10.0          # via plotly
six==1.10.0               # via plotly
